Add files via upload
We have added extra files which are needed by R to run the Stan code and the BYM2 model. 
We also added an illustration of the Stata do file that we used to create the data, and also an R file Script1... which cleans and selects rows.
We will explain more in the actual workshop.  Have a look!
Have fun!  - Wendy Olsen